+++ examples/music/platformio.ini
@@ -14,6 +14,7 @@ default_envs = m5stack-cores3
 [env]
 lib_deps = 
   m5stack/M5Unified
+  ESP32Sampler=symlink://../../
 build_flags =
   -w ;Disable enumeration warnings
