+++ examples/music/src/MidiMessage.h
