+++ examples/music/src/sample/bass.c
