+++ examples/music/src/sample/crash.c
