+++ examples/music/src/sample/hihat.c
