+++ examples/music/src/sample/kick.c
