+++ examples/music/src/sample/rimknock.c
