+++ examples/music/src/sample/snare.c
