+++ examples/music/src/song/simple.cpp
