+++ examples/music/src/song/stresstest.cpp
