+++ include/EffectBase.h
