+++ include/SamplerOptimized.h
@@ -3,41 +3,54 @@
 #include <list>
 #include <deque>
 #include <SamplerBase.h>
-#include <M5Unified.h>
-#include <ml_reverb.h>
+#include <EffectReverb.h>
 
-class SamplerLegacy : public SamplerBase
+class SamplerOptimized : public SamplerBase
 {
 public:
-    struct SamplePlayer
+    // 与えられたサンプルを再生する
+    class SamplePlayer
     {
+    public:
         SamplePlayer(struct Sample *sample, uint8_t noteNo, float volume, float pitchBend)
-            : sample{sample}, noteNo{noteNo}, volume{volume}, pitchBend{pitchBend}, createdAt{micros()} {}
+            : sample{sample}, noteNo{noteNo}, volume{volume}, pitchBend{pitchBend}, createdAt{micros()}
+        {
+            UpdatePitch();
+            gain = volume;
+        }
         SamplePlayer() : sample{nullptr}, noteNo{60}, volume{1.0f}, playing{false}, createdAt{micros()} {}
         struct Sample *sample;
         uint8_t noteNo;
         float pitchBend = 0;
         float volume;
         unsigned long createdAt = 0;
+        bool released = false;
+
+        bool playing = true;
         uint32_t pos = 0;
         float pos_f = 0.0f;
-        bool playing = true;
-        bool released = false;
-        float adsrGain = 0.0f;
+        float gain = 0.0f; // volumeとADSR処理により算出される値
+        float pitch = 1.0f; // noteNoとpitchBendにより算出される値
         enum SampleAdsr adsrState = SampleAdsr::attack;
+
+        void UpdateGain();
+        void UpdatePitch();
+    private:
     };
+
+    // MIDI規格のチャンネルに対応する概念
     class Channel
     {
     public:
         Channel() {}
-        Channel(SamplerLegacy *sampler) : sampler{sampler} {}
+        Channel(SamplerOptimized *sampler) : sampler{sampler} {}
         void NoteOn(uint8_t noteNo, uint8_t velocity);
         void NoteOff(uint8_t noteNo, uint8_t velocity);
         void PitchBend(int16_t pitchBend);
         void SetTimbre(Timbre *timbre);
 
     private:
-        SamplerLegacy *sampler; // 親サンプラー
+        SamplerOptimized *sampler; // 親サンプラー
         struct PlayingNote
         {
             uint8_t noteNo;
@@ -48,18 +61,10 @@ public:
         std::list<PlayingNote> playingNotes; // このチャンネルで現在再生しているノート
     };
 
-    SamplerLegacy()
+    SamplerOptimized()
     {
         for (uint_fast8_t i = 0; i < CH_COUNT; i++)
             channels[i] = Channel(this);
-
-        revBuffer = (float *)heap_caps_aligned_calloc(16, 1, sizeof(float) * REV_BUFF_SIZE, MALLOC_CAP_INTERNAL);
-        Reverb_Setup(revBuffer);
-        Reverb_SetLevel(0, 0.2f);
-    }
-    ~SamplerLegacy()
-    {
-        free(revBuffer);
     }
     void NoteOn(uint8_t noteNo, uint8_t velocity, uint8_t channel);
     void NoteOff(uint8_t noteNo, uint8_t velocity, uint8_t channel);
@@ -71,6 +76,9 @@ public:
     float masterVolume = 0.4f;
 
 private:
+    // 各メッセージのキューイングに使用する
+    // MIDIのメッセージとは互換性がない
+    // TODO: noteNo/velocityとpitchBendは同時に使われることがないのに両方メモリを占有しているのどうにかならないか…？
     struct Message
     {
         uint8_t status;
@@ -87,8 +95,10 @@ private:
     };
     Channel channels[CH_COUNT]; // コンストラクタで初期化する
     SamplePlayer players[MAX_SOUND] = {SamplePlayer()};
+    // 受け取ったNoteOn/NoteOff/PitchBendなどは一旦キューに入れておき、Processのタイミングで処理する
+    // これにより、Processを別スレッドで動かすことができる
+    // TODO: messageQueue自体の排他制御は必要ない？
     std::deque<Message> messageQueue;
-    float *revBuffer;
-    float PitchFromNoteNo(float noteNo, float root, float pitchBend);
-    void UpdateAdsr(SamplePlayer *player);
+
+    EffectReverb reverb = EffectReverb(0.4f, 0.5f, SAMPLE_BUFFER_SIZE, SAMPLE_RATE);
 };
